+++ src/main/java/com/controller/QuestionController.java
@@ -1,7 +1,6 @@
 package com.controller;
 
 import com.dto.QuestionDTO;
-import com.mapper.QuestionMapper;
 import com.service.QuestionService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,7 +16,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id,
+    public String question(@PathVariable(name = "id") Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         //累加阅读数的功能
